pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "An agentic research assistant powered by LangGraph"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "tavily-python>=0.5.0",
    "duckduckgo-search>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.38.0",
    "pydantic>=2.0.0",
]

[project.scripts]
research = "main:main"
